+++ optimizer/__init__.py
@@ -1,3 +1,3 @@
 """Codon Optimizer for Non-Model Organisms."""
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"